Set up the backbone, basic code, not optimised yet.
The Router class communicates directly with the webserver, it takes a
URL, send to the server as a request and get back the JSON text, convert
that text into JSONObject and pass it back to the Renderer. I use an
interface called RequestsListener for event-handling here, and hold a
list of requests listeners in the Renderer class (the origin of the
requests), but am thinking this is quite unnecessary because the
renderer only sends the request to one single router, and the router
only process one request at a time, but this might change later?

Renderer has a method send that triggers all the requests listener (the
router atm). It also has a variable called currentPage that holds the
state of the display, I intend to use this variable to update the screen
whenever the user navigate to a different function on Mobile Oxford